requirements.txt
streamlit
torch
torchvision
opencv-python
Pillow
numpy
scipy
pytesseract